Move `ChannelRegistry` implementation to `_internal/_channels.py`
This is a stand-alone class and users shouldn't be tied to the `actor`
module.
