+++ src/frequenz/sdk/actor/__init__.py
@@ -599,7 +599,6 @@ Received message=False
 from ..timeseries._resampling import ResamplerConfig
 from ._actor import Actor
 from ._background_service import BackgroundService
-from ._channel_registry import ChannelRegistry
 from ._config_managing import ConfigManagingActor
 from ._data_sourcing import ComponentMetricId, ComponentMetricRequest, DataSourcingActor
 from ._resampling import ComponentMetricsResamplingActor
@@ -608,7 +607,6 @@ from ._run_utils import run
 __all__ = [
     "Actor",
     "BackgroundService",
-    "ChannelRegistry",
     "ComponentMetricId",
     "ComponentMetricRequest",
     "ComponentMetricsResamplingActor",
