+++ src/frequenz/sdk/actor/_data_sourcing/microgrid_api_source.py
@@ -18,10 +18,10 @@ from frequenz.client.microgrid import (
     MeterData,
 )
 
+from ..._internal._channels import ChannelRegistry
 from ...microgrid import connection_manager
 from ...timeseries import Sample
 from ...timeseries._quantities import Quantity
-from .._channel_registry import ChannelRegistry
 from ._component_metric_request import ComponentMetricRequest
 
 _logger = logging.getLogger(__name__)
