+++ src/frequenz/sdk/timeseries/_grid_frequency.py
@@ -12,7 +12,7 @@ from typing import TYPE_CHECKING
 from frequenz.channels import Receiver, Sender
 from frequenz.client.microgrid import Component, ComponentCategory, ComponentMetricId
 
-from ..actor import ChannelRegistry
+from .._internal._channels import ChannelRegistry
 from ..microgrid import connection_manager
 from ..timeseries._base_types import Sample
 from ..timeseries._quantities import Frequency, Quantity
