+++ src/frequenz/sdk/timeseries/formula_engine/_formula_generators/_formula_generator.py
@@ -15,6 +15,7 @@ from typing import TYPE_CHECKING, Generic
 from frequenz.channels import Sender
 from frequenz.client.microgrid import Component, ComponentCategory, ComponentMetricId
 
+from ...._internal._channels import ChannelRegistry
 from ....microgrid import connection_manager
 from ..._quantities import QuantityT
 from .._formula_engine import FormulaEngine, FormulaEngine3Phase
@@ -22,7 +23,7 @@ from .._resampled_formula_builder import ResampledFormulaBuilder
 
 if TYPE_CHECKING:
     # Break circular import
-    from ....actor import ChannelRegistry, ComponentMetricRequest
+    from ....actor import ComponentMetricRequest
 
 
 class FormulaGenerationError(Exception):
