+++ src/frequenz/sdk/timeseries/logical_meter/_logical_meter.py
@@ -8,7 +8,8 @@ import uuid
 
 from frequenz.channels import Sender
 
-from ...actor import ChannelRegistry, ComponentMetricId, ComponentMetricRequest
+from ..._internal._channels import ChannelRegistry
+from ...actor import ComponentMetricId, ComponentMetricRequest
 from .._quantities import Power, Quantity
 from ..formula_engine import FormulaEngine
 from ..formula_engine._formula_engine_pool import FormulaEnginePool
